# src/main/resources/openapi.yml
openapi: 3.0.0
info:
  title: Sistema de Reservas de Companhias Aéreas
  description: API para gerenciamento de reservas de voos.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Servidor local
paths:
...
  /passageiros/{id}:
    get:
      summary: Obtém um passageiro pelo ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID do passageiro
      responses:
        '200':
          description: Detalhes do passageiro
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Passenger'
    put:
    patch:
      summary: Atualiza um passageiro pelo ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID do passageiro
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Passenger'
      responses:
        '200':
          description: Passageiro atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Passenger'
    delete:
      summary: Remove um passageiro pelo ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID do passageiro
      responses:
        '204':
          description: Passageiro removido
...
components:
  schemas:
    Passenger:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
...